AuthorizationMicroservice/src/main/java/com/example/authorization/exception/CustomExceptionHandler.java
package com.example.authorization.exception;

import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.ControllerAdvice;

@ControllerAdvice
public class CustomExceptionHandler {
    @org.springframework.web.bind.annotation.ExceptionHandler(AuthException.class)
    public ResponseEntity<String> handleCustomException(AuthException ex) {
        return new ResponseEntity<>(ex.getMessage(), HttpStatus.INTERNAL_SERVER_ERROR);
    }
}
